+++ src/mapfile_parser/__init__.py
@@ -1,11 +1,11 @@
 #!/usr/bin/env python3
 
-# SPDX-FileCopyrightText: © 2022-2023 Decompollaborate
+# SPDX-FileCopyrightText: © 2022-2024 Decompollaborate
 # SPDX-License-Identifier: MIT
 
 from __future__ import annotations
 
-__version_info__ = (2, 3, 1)
+__version_info__ = (2, 3, 2)
 __version__ = ".".join(map(str, __version_info__))
 __author__ = "Decompollaborate"