+++ src/rs/progress_stats.rs
@@ -1,4 +1,4 @@
-/* SPDX-FileCopyrightText: © 2023 Decompollaborate */
+/* SPDX-FileCopyrightText: © 2023-2024 Decompollaborate */
 /* SPDX-License-Identifier: MIT */
 
 use std::collections::HashMap;
@@ -51,31 +51,45 @@ impl ProgressStats {
         categories
     }
 
-    pub fn get_header_as_str() -> String {
+    pub fn get_header_as_str(category_column_size: usize) -> String {
         format!(
-            "{:<28}: {:>12} / {:>8} {:>10}%  ({:>20}%)",
-            "Category", "DecompedSize", "Total", "OfFolder", "OfTotal"
+            "{:<category_column_size$}: {:>12} / {:>8} {:>10}%  ({:>20}%)",
+            "Category",
+            "DecompedSize",
+            "Total",
+            "OfFolder",
+            "OfTotal",
+            category_column_size = category_column_size
         )
     }
 
-    pub fn print_header() {
-        println!("{}", Self::get_header_as_str());
+    pub fn print_header(category_column_size: usize) {
+        println!("{}", Self::get_header_as_str(category_column_size));
     }
 
-    pub fn get_entry_as_str(&self, category: &str, total_stats: &Self) -> String {
+    pub fn get_entry_as_str(
+        &self,
+        category: &str,
+        total_stats: &Self,
+        category_column_size: usize,
+    ) -> String {
         format!(
-            "{:<28}: {:>12} / {:>8} {:>10.4}%  ({:>8.4}% / {:>8.4}%)",
+            "{:<category_column_size$}: {:>12} / {:>8} {:>10.4}%  ({:>8.4}% / {:>8.4}%)",
             category,
             self.decomped_size,
             self.total(),
             self.decomped_percentage(),
             self.decomped_percentage_total(total_stats),
-            self.total() as f32 / total_stats.total() as f32 * 100.0
+            self.total() as f32 / total_stats.total() as f32 * 100.0,
+            category_column_size = category_column_size
         )
     }
 
-    pub fn print(&self, category: &str, total_stats: &Self) {
-        println!("{}", self.get_entry_as_str(category, total_stats));
+    pub fn print(&self, category: &str, total_stats: &Self, category_column_size: usize) {
+        println!(
+            "{}",
+            self.get_entry_as_str(category, total_stats, category_column_size)
+        );
     }
 }
 
@@ -152,22 +166,31 @@ pub(crate) mod python_bindings {
         }
 
         #[staticmethod]
-        pub fn getHeaderAsStr() -> String {
-            Self::get_header_as_str()
+        #[pyo3(signature=(category_column_size=28))]
+        pub fn getHeaderAsStr(category_column_size: usize) -> String {
+            Self::get_header_as_str(category_column_size)
         }
 
         #[staticmethod]
-        pub fn printHeader() {
-            Self::print_header()
+        #[pyo3(signature=(category_column_size=28))]
+        pub fn printHeader(category_column_size: usize) {
+            Self::print_header(category_column_size)
         }
 
-        pub fn getEntryAsStr(&self, category: &str, total_stats: &Self) -> String {
-            self.get_entry_as_str(category, total_stats)
+        #[pyo3(signature=(category, total_stats, category_column_size=28))]
+        pub fn getEntryAsStr(
+            &self,
+            category: &str,
+            total_stats: &Self,
+            category_column_size: usize,
+        ) -> String {
+            self.get_entry_as_str(category, total_stats, category_column_size)
         }
 
         #[pyo3(name = "print")]
-        pub fn py_print(&self, category: &str, total_stats: &Self) {
-            self.print(category, total_stats)
+        #[pyo3(signature=(category, total_stats, category_column_size=28))]
+        pub fn py_print(&self, category: &str, total_stats: &Self, category_column_size: usize) {
+            self.print(category, total_stats, category_column_size)
         }
     }
 }